Update Emails.py
The current bug is a logic problem. The above loop is iterating through and printing all the emails and then it sends only the last one
This is because the loop is running and then after the loop is over it sends the messages. I need it to run the loop, send the message and
then run the loop on the next row